[Reportings] can ope reporting form in side window
