+++ frontend/src/domain/use_cases/reportings/switchReporting.ts
@@ -1,40 +1,44 @@
-import { getIdTyped } from '../../../utils/getIdTyped'
+import { SideWindowReportingFormVisibility } from '../../../features/Reportings/sideWindowContext/context'
 import { setReportingFormVisibility } from '../../shared_slices/Global'
 import { multiReportingsActions } from '../../shared_slices/MultiReportings'
 import { ReportingFormVisibility } from '../../shared_slices/ReportingState'
 
-export const switchReporting = nextReportingId => async (dispatch, getState) => {
-  const {
-    multiReportings: { selectedReportings },
-    reportingState: { context, isFormDirty, reportingState }
-  } = getState()
+export const switchReporting =
+  (nextReportingId, setContextVisibility?: (visiblity) => void) => async (dispatch, getState) => {
+    const {
+      multiReportings: { selectedReportings },
+      reportingState: { context, isFormDirty, reportingState }
+    } = getState()
 
-  const id = getIdTyped(nextReportingId)
+    const updatedReportings = [...selectedReportings]
 
-  const updatedReportings = [...selectedReportings]
-  const reportingIndex = updatedReportings.findIndex(reporting =>
-    reportingState ? reporting.reporting.id === reportingState?.id : reporting.reporting.id === id
-  )
-
-  // We want to save the active form before switching on another
-  if (reportingState && selectedReportings.length > 0) {
-    const formattedReporting = {
-      context,
-      isFormDirty,
-      reporting: reportingState
+    // We want to save the active form before switching on another
+    if (reportingState) {
+      const reportingToSaveIndex = updatedReportings.findIndex(
+        reporting => reportingState && reporting.reporting.id === reportingState?.id
+      )
+      const formattedReporting = {
+        context,
+        isFormDirty,
+        reporting: reportingState
+      }
+      if (reportingToSaveIndex !== -1) {
+        updatedReportings[reportingToSaveIndex] = formattedReporting
+      } else {
+        updatedReportings.push(formattedReporting)
+      }
     }
-    if (reportingIndex !== -1) {
-      updatedReportings[reportingIndex] = formattedReporting
+
+    await dispatch(
+      multiReportingsActions.setSelectedReportings({
+        activeReportingId: nextReportingId,
+        selectedReportings: updatedReportings
+      })
+    )
+
+    if (setContextVisibility) {
+      setContextVisibility(SideWindowReportingFormVisibility.VISIBLE)
     } else {
-      updatedReportings.push(formattedReporting)
+      dispatch(setReportingFormVisibility(ReportingFormVisibility.VISIBLE))
     }
   }
-
-  await dispatch(
-    multiReportingsActions.setSelectedReportings({
-      activeReportingId: nextReportingId,
-      selectedReportings: updatedReportings
-    })
-  )
-  dispatch(setReportingFormVisibility(ReportingFormVisibility.VISIBLE))
-}
