+++ frontend/src/features/Reportings/ReportingForm/FormComponents/Position/PointPicker.tsx
@@ -11,13 +11,13 @@ import {
   OLGeometryType,
   OPENLAYERS_PROJECTION,
   WSG84_PROJECTION
-} from '../../../../domain/entities/map/constants'
-import { setFitToExtent } from '../../../../domain/shared_slices/Map'
-import { drawPoint } from '../../../../domain/use_cases/draw/drawGeometry'
-import { useAppDispatch } from '../../../../hooks/useAppDispatch'
-import { useAppSelector } from '../../../../hooks/useAppSelector'
-import { useListenForDrawedGeometry } from '../../../../hooks/useListenForDrawing'
-import { formatCoordinates } from '../../../../utils/coordinates'
+} from '../../../../../domain/entities/map/constants'
+import { setFitToExtent } from '../../../../../domain/shared_slices/Map'
+import { drawPoint } from '../../../../../domain/use_cases/draw/drawGeometry'
+import { useAppDispatch } from '../../../../../hooks/useAppDispatch'
+import { useAppSelector } from '../../../../../hooks/useAppSelector'
+import { useListenForDrawedGeometry } from '../../../../../hooks/useListenForDrawing'
+import { formatCoordinates } from '../../../../../utils/coordinates'
 
 import type { Coordinate } from 'ol/coordinate'
