+++ frontend/src/features/Reportings/ReportingForm/FormComponents/Source.tsx
@@ -17,13 +17,13 @@ import { GeoJSON } from 'ol/format'
 import { useMemo } from 'react'
 import styled from 'styled-components'
 
-import { useGetControlUnitsQuery } from '../../../api/controlUnitsAPI'
-import { useGetSemaphoresQuery } from '../../../api/semaphoresAPI'
-import { Reporting, ReportingSourceEnum, ReportingSourceLabels } from '../../../domain/entities/reporting'
-import { setDisplayedItems } from '../../../domain/shared_slices/Global'
-import { setZoomToCenter } from '../../../domain/shared_slices/Map'
-import { setIsSemaphoreHighlighted, setSelectedSemaphore } from '../../../domain/shared_slices/SemaphoresSlice'
-import { useAppDispatch } from '../../../hooks/useAppDispatch'
+import { useGetControlUnitsQuery } from '../../../../api/controlUnitsAPI'
+import { useGetSemaphoresQuery } from '../../../../api/semaphoresAPI'
+import { Reporting, ReportingSourceEnum, ReportingSourceLabels } from '../../../../domain/entities/reporting'
+import { setDisplayedItems } from '../../../../domain/shared_slices/Global'
+import { setZoomToCenter } from '../../../../domain/shared_slices/Map'
+import { setIsSemaphoreHighlighted, setSelectedSemaphore } from '../../../../domain/shared_slices/SemaphoresSlice'
+import { useAppDispatch } from '../../../../hooks/useAppDispatch'
 
 import type { Point } from 'ol/geom'
