+++ frontend/src/features/Reportings/ReportingsList/index.tsx
@@ -1,4 +1,5 @@
 import { Button, Icon } from '@mtes-mct/monitor-ui'
+import { useContext } from 'react'
 import styled from 'styled-components'
 
 import { ReportingsTable } from './ReportingsTable'
@@ -7,14 +8,16 @@ import { addReporting } from '../../../domain/use_cases/reportings/addReporting'
 import { useAppDispatch } from '../../../hooks/useAppDispatch'
 import { ReportingsFilters } from '../Filters'
 import { useGetFilteredReportingsQuery } from '../hooks/useGetFilteredReportingsQuery'
+import { SideWindowReportingsContext } from '../sideWindowContext/context'
 
 export function ReportingsList() {
   const dispatch = useAppDispatch()
+  const reportingFormVisibility = useContext(SideWindowReportingsContext)
 
   const { isError, isFetching, isLoading, reportings } = useGetFilteredReportingsQuery()
 
   const createReporting = () => {
-    dispatch(addReporting(ReportingContext.SIDE_WINDOW))
+    dispatch(addReporting(ReportingContext.SIDE_WINDOW, undefined, reportingFormVisibility.setContextVisibility))
   }
 
   return (
