+++ frontend/src/features/SideWindow/index.tsx
@@ -1,5 +1,5 @@
 import { Icon, SideMenu, type NewWindowContextValue, NewWindowContext } from '@mtes-mct/monitor-ui'
-import { forwardRef, useEffect, useImperativeHandle, useMemo, useState, useRef } from 'react'
+import { forwardRef, useEffect, useImperativeHandle, useMemo, useState, useRef, useCallback } from 'react'
 import { useDispatch } from 'react-redux'
 import { generatePath } from 'react-router'
 import { ToastContainer } from 'react-toastify'
@@ -15,6 +15,8 @@ import { Mission } from '../missions/MissionForm'
 import { Missions } from '../missions/MissionsList'
 import { MissionsNavBar } from '../missions/MissionsNavBar'
 import { ReportingsList } from '../Reportings/ReportingsList'
+import { ReportingsOnSideWindow } from '../Reportings/sideWindowContext'
+import { SideWindowReportingFormVisibility, SideWindowReportingsContext } from '../Reportings/sideWindowContext/context'
 
 import type { ForwardedRef, MutableRefObject } from 'react'
 
@@ -27,6 +29,20 @@ function SideWindowWithRef(_, ref: ForwardedRef<HTMLDivElement | null>) {
   const [isFirstRender, setIsFirstRender] = useState(true)
 
   const dispatch = useDispatch()
+
+  const [visibility, setVisibility] = useState(SideWindowReportingFormVisibility.NONE)
+
+  const setContextVisibility = useCallback(nextVisibility => {
+    setVisibility(nextVisibility)
+  }, [])
+
+  const contextValue = useMemo(
+    () => ({
+      contextVisibility: visibility,
+      setContextVisibility
+    }),
+    [setContextVisibility, visibility]
+  )
 
   const isMissionButtonIsActive = useMemo(() => isMissionOrMissionsPage(currentPath), [currentPath])
   const isReportingsButtonIsActive = useMemo(() => isReportingsPage(currentPath), [currentPath])
@@ -55,34 +71,37 @@ function SideWindowWithRef(_, ref: ForwardedRef<HTMLDivElement | null>) {
 
   return (
     <ErrorBoundary>
-      <Wrapper ref={wrapperRef}>
-        {!isFirstRender && (
-          <NewWindowContext.Provider value={newWindowContextProviderValue}>
-            <SideMenu>
-              <SideMenu.Button
-                Icon={Icon.MissionAction}
-                isActive={isMissionButtonIsActive}
-                onClick={() => navigate(generatePath(sideWindowPaths.MISSIONS))}
-                title="missions"
-              />
-              <SideMenu.Button
-                Icon={Icon.Report}
-                isActive={isReportingsButtonIsActive}
-                onClick={() => navigate(generatePath(sideWindowPaths.REPORTINGS))}
-                title="signalements"
-              />
-            </SideMenu>
+      <SideWindowReportingsContext.Provider value={contextValue}>
+        <Wrapper ref={wrapperRef}>
+          {!isFirstRender && (
+            <NewWindowContext.Provider value={newWindowContextProviderValue}>
+              <SideMenu>
+                <SideMenu.Button
+                  Icon={Icon.MissionAction}
+                  isActive={isMissionButtonIsActive}
+                  onClick={() => navigate(generatePath(sideWindowPaths.MISSIONS))}
+                  title="missions"
+                />
+                <SideMenu.Button
+                  Icon={Icon.Report}
+                  isActive={isReportingsButtonIsActive}
+                  onClick={() => navigate(generatePath(sideWindowPaths.REPORTINGS))}
+                  title="signalements"
+                />
+              </SideMenu>
 
-            <StyledRouteContainer>
-              <Route element={<ReportingsList />} path={sideWindowPaths.REPORTINGS} />
-              <Route element={<MissionsNavBar />} path={[sideWindowPaths.MISSIONS, sideWindowPaths.MISSION]} />
-              <Route element={<Missions />} path={sideWindowPaths.MISSIONS} />
-              <Route element={<Mission />} path={sideWindowPaths.MISSION} />
-            </StyledRouteContainer>
-          </NewWindowContext.Provider>
-        )}
-        <ToastContainer containerId="sideWindow" enableMultiContainer />
-      </Wrapper>
+              <StyledRouteContainer>
+                <Route element={<ReportingsList />} path={sideWindowPaths.REPORTINGS} />
+                <Route element={<MissionsNavBar />} path={[sideWindowPaths.MISSIONS, sideWindowPaths.MISSION]} />
+                <Route element={<Missions />} path={sideWindowPaths.MISSIONS} />
+                <Route element={<Mission />} path={sideWindowPaths.MISSION} />
+              </StyledRouteContainer>
+            </NewWindowContext.Provider>
+          )}
+          <ReportingsOnSideWindow />
+          <ToastContainer containerId="sideWindow" enableMultiContainer />
+        </Wrapper>
+      </SideWindowReportingsContext.Provider>
     </ErrorBoundary>
   )
 }
